move changes into gemma/hlo.py
